requirements.txt
fastapi==0.78.0
fastapi-utils==0.2.1
pydantic==1.9.0
uvicorn==0.17.6
python-dotenv==0.20.0
databases==0.5.5
asyncpg==0.25.0
psycopg2==2.9.3
toolz==0.11.2
alembic==1.7.6
pytest==7.1.2
passlib==1.7.4
argon2-cffi==21.3.0
email-validator==1.2.1
pytest-asyncio==0.18.2
setuptools==62.3.2
